move extended layer config to config/
